Move methods on main.py to methods.py
